Refactored Inventory page into bakedGoods, added functionality to baked goods page to display baked goods and quantity from database.
