+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import 'pages/registration.dart'; // Import the RegistrationPage
 import 'pages/registerEmailAndPhone.dart'; // Import the RegisterEmailAndPhone Page
 import 'pages/login.dart';        // Import the LoginPage
-import 'pages/inventoryPage.dart'; //Import the Inventory Page
+import 'pages/bakedGoodsPage.dart'; //Import the Inventory Page
 import 'services/session_service.dart';
 import 'pages/recipePage.dart'; // Import the Recipes Page
 import 'pages/ingredientsPage.dart'; // Import the Ingredients Page
@@ -304,7 +304,7 @@ class HomePage extends StatelessWidget {
                   'Baked Goods',
                   Icons.inventory,
                   onTap: () {
-                    onPageTap(InventoryDetailPage());
+                    onPageTap(BakedGoodsDetailPage());
                   },
                 ),
                 _buildRoundedButton(
