+++ lib/models/bakedGoods.dart
@@ -1,0 +1,7 @@
+class BakedGoods {
+  final String id;
+  final String name;
+  final int quantity;
+
+  BakedGoods({required this.id, required this.name, required this.quantity});
+}
